+++ cache.h
@@ -8,14 +8,14 @@
 #include "util.h"
 #include "packet.h"
 
-void update_cache(Packet_t **cache, unsigned int cache_len);
+void update_cache(Packet_t **cache, size_t cache_len);
 
-void put_in_cache(Packet_t **cache, unsigned int cache_len, Packet_t *packet);
+void put_in_cache(Packet_t **cache, size_t cache_len, Packet_t *packet);
 
-Packet_t *find_in_cache(Packet_t **cache, unsigned int cache_len, char *QNAME);
+Packet_t *find_in_cache(Packet_t **cache, size_t cache_len, char *QNAME);
 
-unsigned int num_empty_slots(Packet_t **cache, unsigned int cache_len);
+size_t num_empty_slots(Packet_t **cache, size_t cache_len);
 
-void reformat_cache(Packet_t **cache, unsigned int cache_len);
+void reformat_cache(Packet_t **cache, size_t cache_len);
 
 #endif
